requirements.txt
flask
pandas
requests
python-dotenv
selenium
beautifulsoup4